refactor: moved mapAesthetics to more logical spot
